#19 Reduce ResultSet extenders to ExtendedResultSet
