+++ CHANGES.md
@@ -6,3 +6,4 @@
 *   [[#3](../../issues/3)] BoxedResultSEt
 *   [[#4](../../issues/4)] OptionalResultSet
 *   [[#6](../../issues/6)] ExtendedRowMapper (BoxedRowMapper, OptionalRowMapper)
+*   [[#19](../../issues/19)] Reduce ResultSet extenders(BoxedRowMapper, OptionalRowMapper) to ExtendedResultSet
