+++ src/main/java/ru/spb/devclub/utils/db/ExtendedRowMapper.java
@@ -2,6 +2,7 @@ package ru.spb.devclub.utils.db;
 
 import org.springframework.jdbc.core.RowMapper;
 
+import java.sql.ResultSet;
 import java.sql.SQLException;
 
 /**
@@ -13,6 +14,11 @@ import java.sql.SQLException;
  * @see ru.spb.devclub.utils.db.ExtendedResultSet
  * @since 0.1.0
  */
-public interface ExtendedRowMapper<T, R extends ExtendedResultSet> extends RowMapper<T> {
-    T mapRow(R rs, int rowNum) throws SQLException;
+public interface ExtendedRowMapper<T> extends RowMapper<T> {
+    T mapRow(ExtendedResultSet rs, int rowNum) throws SQLException;
+
+    @Override
+    default T mapRow(ResultSet rs, int rowNum) throws SQLException {
+        return mapRow(ResultSetUtils.extended(rs), rowNum);
+    }
 }
