+++ src/test/java/ru/spb/devclub/utils/db/SimpleExtendedResultSetTest.java
@@ -6,24 +6,238 @@ import com.mockrunner.mock.jdbc.MockClob;
 import com.mockrunner.mock.jdbc.MockRef;
 import org.junit.Before;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.Mockito;
-import org.mockito.MockitoAnnotations;
+import org.mockito.junit.MockitoJUnitRunner;
 
 import java.math.BigDecimal;
 import java.sql.*;
 import java.util.Optional;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.*;
 
-public class SimpleOptionalResultSetTest {
+@RunWith(MockitoJUnitRunner.class)
+public class SimpleExtendedResultSetTest {
+    private static String columnName = "column";
+    private static int columnIndex = 0;
     @Mock
-    private ResultSet mockResultSet;
+    protected ResultSet mockResultSet;
+    protected ExtendedResultSet extendedResultSet;
 
     @Before
     public void setUp() {
-        MockitoAnnotations.initMocks(this);
+        extendedResultSet = ResultSetUtils.extended(mockResultSet);
+    }
+
+    @Test
+    public void checkCorrectValueFromGetLongOrNullByColumnName() throws SQLException {
+        Long expected = 1L;
+        Mockito.when(mockResultSet.getLong(columnName)).thenReturn(expected);
+
+        assertEquals(expected, extendedResultSet.getLongOrNull(columnName));
+    }
+
+    @Test
+    public void checkNullValueFromGetLongOrNullByColumnName() throws SQLException {
+        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
+
+        assertNull(extendedResultSet.getLongOrNull(columnName));
+    }
+
+    @Test
+    public void checkCorrectValueFromGetLongOrNullByColumnIndex() throws SQLException {
+        Long expected = 1L;
+        Mockito.when(mockResultSet.getLong(columnIndex)).thenReturn(expected);
+
+        assertEquals(expected, extendedResultSet.getLongOrNull(columnIndex));
+    }
+
+    @Test
+    public void checkNullValueFromGetLongOrNullByColumnIndex() throws SQLException {
+        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
+
+        assertNull(extendedResultSet.getLongOrNull(columnIndex));
+    }
+
+    @Test
+    public void checkCorrectValueFromGetIntOrNullByColumnName() throws SQLException {
+        Integer expected = 1;
+        Mockito.when(mockResultSet.getInt(columnName)).thenReturn(expected);
+
+        assertEquals(expected, extendedResultSet.getIntOrNull(columnName));
+    }
+
+    @Test
+    public void checkNullValueFromGetIntOrNullByColumnName() throws SQLException {
+        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
+
+        assertNull(extendedResultSet.getIntOrNull(columnName));
+    }
+
+    @Test
+    public void checkCorrectValueFromGetIntOrNullByColumnIndex() throws SQLException {
+        Integer expected = 1;
+        Mockito.when(mockResultSet.getInt(columnIndex)).thenReturn(expected);
+
+        assertEquals(expected, extendedResultSet.getIntOrNull(columnIndex));
+    }
+
+    @Test
+    public void checkNullValueFromGetIntOrNullByColumnIndex() throws SQLException {
+        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
+
+        assertNull(extendedResultSet.getIntOrNull(columnIndex));
+    }
+
+    @Test
+    public void checkCorrectValueFromGetBooleanOrNullByColumnName() throws SQLException {
+        Boolean expected = true;
+        Mockito.when(mockResultSet.getBoolean(columnName)).thenReturn(expected);
+
+        assertEquals(expected, extendedResultSet.getBooleanOrNull(columnName));
+    }
+
+    @Test
+    public void checkNullValueFromGetBooleanOrNullByColumnName() throws SQLException {
+        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
+
+        assertNull(extendedResultSet.getBooleanOrNull(columnName));
+    }
+
+    @Test
+    public void checkCorrectValueFromGetBooleanOrNullByColumnIndex() throws SQLException {
+        Boolean expected = true;
+        Mockito.when(mockResultSet.getBoolean(columnIndex)).thenReturn(expected);
+
+        assertEquals(expected, extendedResultSet.getBooleanOrNull(columnIndex));
+    }
+
+    @Test
+    public void checkNullValueFromGetBooleanOrNullByColumnIndex() throws SQLException {
+        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
+
+        assertNull(extendedResultSet.getBooleanOrNull(columnIndex));
+    }
+
+    @Test
+    public void checkCorrectValueFromGetShortOrNullByColumnName() throws SQLException {
+        Short expected = 1;
+        Mockito.when(mockResultSet.getShort(columnName)).thenReturn(expected);
+
+        assertEquals(expected, extendedResultSet.getShortOrNull(columnName));
+    }
+
+    @Test
+    public void checkNullValueFromGetShortOrNullByColumnName() throws SQLException {
+        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
+
+        assertNull(extendedResultSet.getShortOrNull(columnName));
+    }
+
+    @Test
+    public void checkCorrectValueFromGetShortOrNullByColumnIndex() throws SQLException {
+        Short expected = 1;
+        Mockito.when(mockResultSet.getShort(columnIndex)).thenReturn(expected);
+
+        assertEquals(expected, extendedResultSet.getShortOrNull(columnIndex));
+    }
+
+    @Test
+    public void checkNullValueFromGetShortOrNullByColumnIndex() throws SQLException {
+        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
+
+        assertNull(extendedResultSet.getShortOrNull(columnIndex));
+    }
+
+    @Test
+    public void checkCorrectValueFromGetByteOrNullByColumnName() throws SQLException {
+        Byte expected = 1;
+        Mockito.when(mockResultSet.getByte(columnName)).thenReturn(expected);
+
+        assertEquals(expected, extendedResultSet.getByteOrNull(columnName));
+    }
+
+    @Test
+    public void checkNullValueFromGetByteOrNullByColumnName() throws SQLException {
+        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
+
+        assertNull(extendedResultSet.getByteOrNull(columnName));
+    }
+
+    @Test
+    public void checkCorrectValueFromGetByteOrNullByColumnIndex() throws SQLException {
+        Byte expected = 1;
+        Mockito.when(mockResultSet.getByte(columnIndex)).thenReturn(expected);
+
+        assertEquals(expected, extendedResultSet.getByteOrNull(columnIndex));
+    }
+
+    @Test
+    public void checkNullValueFromGetByteOrNullByColumnIndex() throws SQLException {
+        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
+
+        assertNull(extendedResultSet.getByteOrNull(columnIndex));
+    }
+
+    @Test
+    public void checkCorrectValueFromGetDoubleOrNullByColumnName() throws SQLException {
+        Double expected = 1.;
+        Mockito.when(mockResultSet.getDouble(columnName)).thenReturn(expected);
+
+        assertEquals(expected, extendedResultSet.getDoubleOrNull(columnName));
+    }
+
+    @Test
+    public void checkNullValueFromGetDoubleOrNullByColumnName() throws SQLException {
+        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
+
+        assertNull(extendedResultSet.getDoubleOrNull(columnName));
+    }
+
+    @Test
+    public void checkCorrectValueFromGetDoubleOrNullByColumnIndex() throws SQLException {
+        Double expected = 1.;
+        Mockito.when(mockResultSet.getDouble(columnIndex)).thenReturn(expected);
+
+        assertEquals(expected, extendedResultSet.getDoubleOrNull(columnIndex));
+    }
+
+    @Test
+    public void checkNullValueFromGetDoubleOrNullByColumnIndex() throws SQLException {
+        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
+
+        assertNull(extendedResultSet.getDoubleOrNull(columnIndex));
+    }
+
+    @Test
+    public void checkCorrectValueFromGetFloatOrNullByColumnName() throws SQLException {
+        Float expected = 1.f;
+        Mockito.when(mockResultSet.getFloat(columnName)).thenReturn(expected);
+
+        assertEquals(expected, extendedResultSet.getFloatOrNull(columnName));
+    }
+
+    @Test
+    public void checkNullValueFromGetFloatOrNullByColumnName() throws SQLException {
+        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
+
+        assertNull(extendedResultSet.getFloatOrNull(columnName));
+    }
+
+    @Test
+    public void checkCorrectValueFromGetFloatOrNullByColumnIndex() throws SQLException {
+        Float expected = 1.f;
+        Mockito.when(mockResultSet.getFloat(columnIndex)).thenReturn(expected);
+
+        assertEquals(expected, extendedResultSet.getFloatOrNull(columnIndex));
+    }
+
+    @Test
+    public void checkNullValueFromGetFloatOrNullByColumnIndex() throws SQLException {
+        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
+
+        assertNull(extendedResultSet.getFloatOrNull(columnIndex));
     }
 
     @Test
@@ -31,17 +245,19 @@ public class SimpleOptionalResultSetTest {
         int columnIndex = 0;
         String expected = "forty thousand";
         Mockito.when(mockResultSet.getString(columnIndex)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<String> optionalString = resultSet.getOptionalString(columnIndex);
+        
+        Optional<String> optionalString = extendedResultSet.getOptionalString(columnIndex);
+        
         assertEquals(expected, optionalString.orElseThrow(IllegalArgumentException::new));
     }
 
     @Test
     public void checkNullFromGetOptionalStringByColumnIndex() throws SQLException {
         int columnIndex = 0;
-        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<String> optionalString = resultSet.getOptionalString(columnIndex);
+        Mockito.when(mockResultSet.getString(columnIndex)).thenReturn(null);
+        
+        Optional<String> optionalString = extendedResultSet.getOptionalString(columnIndex);
+        
         assertFalse(optionalString.isPresent());
     }
 
@@ -50,17 +266,19 @@ public class SimpleOptionalResultSetTest {
         String columnName = "ignored";
         String expected = "forty thousand";
         Mockito.when(mockResultSet.getString(columnName)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<String> optionalString = resultSet.getOptionalString(columnName);
+
+        Optional<String> optionalString = extendedResultSet.getOptionalString(columnName);
+
         assertEquals(expected, optionalString.orElseThrow(IllegalArgumentException::new));
     }
 
     @Test
     public void checkNullFromGetOptionalStringByColumnName() throws SQLException {
         String columnName = "ignored";
-        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<String> optionalString = resultSet.getOptionalString(columnName);
+        Mockito.when(mockResultSet.getString(columnName)).thenReturn(null);
+        
+        Optional<String> optionalString = extendedResultSet.getOptionalString(columnName);
+
         assertFalse(optionalString.isPresent());
     }
 
@@ -69,8 +287,9 @@ public class SimpleOptionalResultSetTest {
         int columnIndex = 0;
         boolean expected = true;
         Mockito.when(mockResultSet.getBoolean(columnIndex)).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Boolean> optionalBoolean = resultSet.getOptionalBoolean(columnIndex);
+        
+        Optional<Boolean> optionalBoolean = extendedResultSet.getOptionalBoolean(columnIndex);
+
         assertEquals(expected, optionalBoolean.orElseThrow(IllegalArgumentException::new));
     }
 
@@ -78,8 +297,9 @@ public class SimpleOptionalResultSetTest {
     public void checkNullFromGetOptionalBooleanByColumnIndex() throws SQLException {
         int columnIndex = 0;
         Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Boolean> optionalBoolean = resultSet.getOptionalBoolean(columnIndex);
+        
+        Optional<Boolean> optionalBoolean = extendedResultSet.getOptionalBoolean(columnIndex);
+
         assertFalse(optionalBoolean.isPresent());
     }
 
@@ -88,8 +308,9 @@ public class SimpleOptionalResultSetTest {
         String columnName = "ignored";
         boolean expected = true;
         Mockito.when(mockResultSet.getBoolean(columnName)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Boolean> optionalBoolean = resultSet.getOptionalBoolean(columnName);
+        
+        Optional<Boolean> optionalBoolean = extendedResultSet.getOptionalBoolean(columnName);
+
         assertEquals(expected, optionalBoolean.orElseThrow(IllegalArgumentException::new));
     }
 
@@ -97,8 +318,9 @@ public class SimpleOptionalResultSetTest {
     public void checkNullFromGetOptionalBooleanByColumnName() throws SQLException {
         String columnName = "ignored";
         Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Boolean> optionalBoolean = resultSet.getOptionalBoolean(columnName);
+        
+        Optional<Boolean> optionalBoolean = extendedResultSet.getOptionalBoolean(columnName);
+
         assertFalse(optionalBoolean.isPresent());
     }
 
@@ -107,8 +329,9 @@ public class SimpleOptionalResultSetTest {
         int columnIndex = 0;
         Byte expected = 40;
         Mockito.when(mockResultSet.getByte(columnIndex)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Byte> optionalByte = resultSet.getOptionalByte(columnIndex);
+        
+        Optional<Byte> optionalByte = extendedResultSet.getOptionalByte(columnIndex);
+
         assertEquals(expected, optionalByte.orElseThrow(IllegalArgumentException::new));
     }
 
@@ -116,8 +339,9 @@ public class SimpleOptionalResultSetTest {
     public void checkNullFromGetOptionalByteByColumnIndex() throws SQLException {
         int columnIndex = 0;
         Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Byte> optionalByte = resultSet.getOptionalByte(columnIndex);
+        
+        Optional<Byte> optionalByte = extendedResultSet.getOptionalByte(columnIndex);
+
         assertFalse(optionalByte.isPresent());
     }
 
@@ -126,8 +350,9 @@ public class SimpleOptionalResultSetTest {
         String columnName = "ignored";
         Byte expected = 40;
         Mockito.when(mockResultSet.getByte(columnName)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Byte> optionalByte = resultSet.getOptionalByte(columnName);
+        
+        Optional<Byte> optionalByte = extendedResultSet.getOptionalByte(columnName);
+
         assertEquals(expected, optionalByte.orElseThrow(IllegalArgumentException::new));
     }
 
@@ -135,8 +360,9 @@ public class SimpleOptionalResultSetTest {
     public void checkNullFromGetOptionalByteByColumnName() throws SQLException {
         String columnName = "ignored";
         Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Byte> optionalByte = resultSet.getOptionalByte(columnName);
+        
+        Optional<Byte> optionalByte = extendedResultSet.getOptionalByte(columnName);
+
         assertFalse(optionalByte.isPresent());
     }
 
@@ -145,8 +371,9 @@ public class SimpleOptionalResultSetTest {
         int columnIndex = 0;
         Short expected = 4_000;
         Mockito.when(mockResultSet.getShort(columnIndex)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Short> optionalShort = resultSet.getOptionalShort(columnIndex);
+        
+        Optional<Short> optionalShort = extendedResultSet.getOptionalShort(columnIndex);
+
         assertEquals(expected, optionalShort.orElseThrow(IllegalArgumentException::new));
     }
 
@@ -154,8 +381,9 @@ public class SimpleOptionalResultSetTest {
     public void checkNullFromGetOptionalShortByColumnIndex() throws SQLException {
         int columnIndex = 0;
         Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Short> optionalShort = resultSet.getOptionalShort(columnIndex);
+        
+        Optional<Short> optionalShort = extendedResultSet.getOptionalShort(columnIndex);
+
         assertFalse(optionalShort.isPresent());
     }
 
@@ -164,8 +392,9 @@ public class SimpleOptionalResultSetTest {
         String columnName = "ignored";
         Short expected = 4_000;
         Mockito.when(mockResultSet.getShort(columnName)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Short> optionalShort = resultSet.getOptionalShort(columnName);
+        
+        Optional<Short> optionalShort = extendedResultSet.getOptionalShort(columnName);
+
         assertEquals(expected, optionalShort.orElseThrow(IllegalArgumentException::new));
     }
 
@@ -173,8 +402,9 @@ public class SimpleOptionalResultSetTest {
     public void checkNullFromGetOptionalShortByColumnName() throws SQLException {
         String columnName = "ignored";
         Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Short> optionalShort = resultSet.getOptionalShort(columnName);
+        
+        Optional<Short> optionalShort = extendedResultSet.getOptionalShort(columnName);
+
         assertFalse(optionalShort.isPresent());
     }
 
@@ -183,8 +413,9 @@ public class SimpleOptionalResultSetTest {
         int columnIndex = 0;
         Integer expected = 40_000;
         Mockito.when(mockResultSet.getInt(columnIndex)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Integer> optionalInteger = resultSet.getOptionalInt(columnIndex);
+        
+        Optional<Integer> optionalInteger = extendedResultSet.getOptionalInt(columnIndex);
+
         assertEquals(expected, optionalInteger.orElseThrow(IllegalArgumentException::new));
     }
 
@@ -192,8 +423,9 @@ public class SimpleOptionalResultSetTest {
     public void checkNullFromGetOptionalIntegerByColumnIndex() throws SQLException {
         int columnIndex = 0;
         Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Integer> optionalInteger = resultSet.getOptionalInt(columnIndex);
+        
+        Optional<Integer> optionalInteger = extendedResultSet.getOptionalInt(columnIndex);
+
         assertFalse(optionalInteger.isPresent());
     }
 
@@ -202,8 +434,9 @@ public class SimpleOptionalResultSetTest {
         String columnName = "ignored";
         Integer expected = 40_000;
         Mockito.when(mockResultSet.getInt(columnName)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Integer> optionalInteger = resultSet.getOptionalInt(columnName);
+        
+        Optional<Integer> optionalInteger = extendedResultSet.getOptionalInt(columnName);
+
         assertEquals(expected, optionalInteger.orElseThrow(IllegalArgumentException::new));
     }
 
@@ -211,8 +444,9 @@ public class SimpleOptionalResultSetTest {
     public void checkNullFromGetOptionalIntegerByColumnName() throws SQLException {
         String columnName = "ignored";
         Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Integer> optionalInteger = resultSet.getOptionalInt(columnName);
+        
+        Optional<Integer> optionalInteger = extendedResultSet.getOptionalInt(columnName);
+
         assertFalse(optionalInteger.isPresent());
     }
 
@@ -221,8 +455,9 @@ public class SimpleOptionalResultSetTest {
         int columnIndex = 0;
         Long expected = 40_000L;
         Mockito.when(mockResultSet.getLong(columnIndex)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Long> optionalLong = resultSet.getOptionalLong(columnIndex);
+        
+        Optional<Long> optionalLong = extendedResultSet.getOptionalLong(columnIndex);
+
         assertEquals(expected, optionalLong.orElseThrow(IllegalArgumentException::new));
     }
 
@@ -230,8 +465,9 @@ public class SimpleOptionalResultSetTest {
     public void checkNullFromGetOptionalLongByColumnIndex() throws SQLException {
         int columnIndex = 0;
         Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Long> optionalLong = resultSet.getOptionalLong(columnIndex);
+        
+        Optional<Long> optionalLong = extendedResultSet.getOptionalLong(columnIndex);
+
         assertFalse(optionalLong.isPresent());
     }
 
@@ -240,8 +476,9 @@ public class SimpleOptionalResultSetTest {
         String columnName = "ignored";
         Long expected = 40_000L;
         Mockito.when(mockResultSet.getLong(columnName)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Long> optionalLong = resultSet.getOptionalLong(columnName);
+        
+        Optional<Long> optionalLong = extendedResultSet.getOptionalLong(columnName);
+
         assertEquals(expected, optionalLong.orElseThrow(IllegalArgumentException::new));
     }
 
@@ -249,8 +486,9 @@ public class SimpleOptionalResultSetTest {
     public void checkNullFromGetOptionalLongByColumnName() throws SQLException {
         String columnName = "ignored";
         Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Long> optionalLong = resultSet.getOptionalLong(columnName);
+        
+        Optional<Long> optionalLong = extendedResultSet.getOptionalLong(columnName);
+
         assertFalse(optionalLong.isPresent());
     }
 
@@ -259,8 +497,9 @@ public class SimpleOptionalResultSetTest {
         int columnIndex = 0;
         Float expected = 40_000F;
         Mockito.when(mockResultSet.getFloat(columnIndex)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Float> optionalFloat = resultSet.getOptionalFloat(columnIndex);
+        
+        Optional<Float> optionalFloat = extendedResultSet.getOptionalFloat(columnIndex);
+
         assertEquals(expected, optionalFloat.orElseThrow(IllegalArgumentException::new));
     }
 
@@ -268,8 +507,9 @@ public class SimpleOptionalResultSetTest {
     public void checkNullFromGetOptionalFloatByColumnIndex() throws SQLException {
         int columnIndex = 0;
         Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Float> optionalFloat = resultSet.getOptionalFloat(columnIndex);
+        
+        Optional<Float> optionalFloat = extendedResultSet.getOptionalFloat(columnIndex);
+
         assertFalse(optionalFloat.isPresent());
     }
 
@@ -278,8 +518,9 @@ public class SimpleOptionalResultSetTest {
         String columnName = "ignored";
         Float expected = 40_000F;
         Mockito.when(mockResultSet.getFloat(columnName)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Float> optionalFloat = resultSet.getOptionalFloat(columnName);
+        
+        Optional<Float> optionalFloat = extendedResultSet.getOptionalFloat(columnName);
+
         assertEquals(expected, optionalFloat.orElseThrow(IllegalArgumentException::new));
     }
 
@@ -287,8 +528,9 @@ public class SimpleOptionalResultSetTest {
     public void checkNullFromGetOptionalFloatByColumnName() throws SQLException {
         String columnName = "ignored";
         Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Float> optionalFloat = resultSet.getOptionalFloat(columnName);
+        
+        Optional<Float> optionalFloat = extendedResultSet.getOptionalFloat(columnName);
+
         assertFalse(optionalFloat.isPresent());
     }
 
@@ -297,8 +539,9 @@ public class SimpleOptionalResultSetTest {
         int columnIndex = 0;
         Double expected = 40_000D;
         Mockito.when(mockResultSet.getDouble(columnIndex)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Double> optionalDouble = resultSet.getOptionalDouble(columnIndex);
+        
+        Optional<Double> optionalDouble = extendedResultSet.getOptionalDouble(columnIndex);
+
         assertEquals(expected, optionalDouble.orElseThrow(IllegalArgumentException::new), .1);
     }
 
@@ -306,8 +549,9 @@ public class SimpleOptionalResultSetTest {
     public void checkNullFromGetOptionalDoubleByColumnIndex() throws SQLException {
         int columnIndex = 0;
         Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Double> optionalDouble = resultSet.getOptionalDouble(columnIndex);
+        
+        Optional<Double> optionalDouble = extendedResultSet.getOptionalDouble(columnIndex);
+
         assertFalse(optionalDouble.isPresent());
     }
 
@@ -316,8 +560,9 @@ public class SimpleOptionalResultSetTest {
         String columnName = "ignored";
         Double expected = 40_000D;
         Mockito.when(mockResultSet.getDouble(columnName)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Double> optionalDouble = resultSet.getOptionalDouble(columnName);
+        
+        Optional<Double> optionalDouble = extendedResultSet.getOptionalDouble(columnName);
+
         assertEquals(expected, optionalDouble.orElseThrow(IllegalArgumentException::new), .1);
     }
 
@@ -325,8 +570,9 @@ public class SimpleOptionalResultSetTest {
     public void checkNullFromGetOptionalDoubleByColumnName() throws SQLException {
         String columnName = "ignored";
         Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Double> optionalDouble = resultSet.getOptionalDouble(columnName);
+        
+        Optional<Double> optionalDouble = extendedResultSet.getOptionalDouble(columnName);
+
         assertFalse(optionalDouble.isPresent());
     }
 
@@ -335,17 +581,19 @@ public class SimpleOptionalResultSetTest {
         int columnIndex = 0;
         byte[] expected = new byte[]{4, 0, 0, 0, 0};
         Mockito.when(mockResultSet.getBytes(columnIndex)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<byte[]> optionalBytes = resultSet.getOptionalBytes(columnIndex);
+        
+        Optional<byte[]> optionalBytes = extendedResultSet.getOptionalBytes(columnIndex);
+
         assertEquals(expected, optionalBytes.orElseThrow(IllegalArgumentException::new));
     }
 
     @Test
     public void checkNullFromGetOptionalBytesByColumnIndex() throws SQLException {
         int columnIndex = 0;
-        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<byte[]> optionalBytes = resultSet.getOptionalBytes(columnIndex);
+        Mockito.when(mockResultSet.getBytes(columnIndex)).thenReturn(null);
+        
+        Optional<byte[]> optionalBytes = extendedResultSet.getOptionalBytes(columnIndex);
+
         assertFalse(optionalBytes.isPresent());
     }
 
@@ -354,17 +602,19 @@ public class SimpleOptionalResultSetTest {
         String columnName = "ignored";
         byte[] expected = new byte[]{4, 0, 0, 0, 0};
         Mockito.when(mockResultSet.getBytes(columnName)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<byte[]> optionalBytes = resultSet.getOptionalBytes(columnName);
+        
+        Optional<byte[]> optionalBytes = extendedResultSet.getOptionalBytes(columnName);
+
         assertEquals(expected, optionalBytes.orElseThrow(IllegalArgumentException::new));
     }
 
     @Test
     public void checkNullFromGetOptionalBytesByColumnName() throws SQLException {
         String columnName = "ignored";
-        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<byte[]> optionalBytes = resultSet.getOptionalBytes(columnName);
+        Mockito.when(mockResultSet.getBytes(columnName)).thenReturn(null);
+        
+        Optional<byte[]> optionalBytes = extendedResultSet.getOptionalBytes(columnName);
+
         assertFalse(optionalBytes.isPresent());
     }
 
@@ -373,17 +623,19 @@ public class SimpleOptionalResultSetTest {
         int columnIndex = 0;
         Date expected = Date.valueOf("2004-04-04");
         Mockito.when(mockResultSet.getDate(columnIndex)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Date> optionalDate = resultSet.getOptionalDate(columnIndex);
+        
+        Optional<Date> optionalDate = extendedResultSet.getOptionalDate(columnIndex);
+
         assertEquals(expected, optionalDate.orElseThrow(IllegalArgumentException::new));
     }
 
     @Test
     public void checkNullFromGetOptionalDateByColumnIndex() throws SQLException {
         int columnIndex = 0;
-        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Date> optionalDate = resultSet.getOptionalDate(columnIndex);
+        Mockito.when(mockResultSet.getDate(columnIndex)).thenReturn(null);
+        
+        Optional<Date> optionalDate = extendedResultSet.getOptionalDate(columnIndex);
+
         assertFalse(optionalDate.isPresent());
     }
 
@@ -392,17 +644,19 @@ public class SimpleOptionalResultSetTest {
         String columnName = "ignored";
         Date expected = Date.valueOf("2004-04-04");
         Mockito.when(mockResultSet.getDate(columnName)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Date> optionalDate = resultSet.getOptionalDate(columnName);
+        
+        Optional<Date> optionalDate = extendedResultSet.getOptionalDate(columnName);
+
         assertEquals(expected, optionalDate.orElseThrow(IllegalArgumentException::new));
     }
 
     @Test
     public void checkNullFromGetOptionalDateByColumnName() throws SQLException {
         String columnName = "ignored";
-        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Date> optionalDate = resultSet.getOptionalDate(columnName);
+        Mockito.when(mockResultSet.getDate(columnName)).thenReturn(null);
+        
+        Optional<Date> optionalDate = extendedResultSet.getOptionalDate(columnName);
+
         assertFalse(optionalDate.isPresent());
     }
 
@@ -411,17 +665,19 @@ public class SimpleOptionalResultSetTest {
         int columnIndex = 0;
         Time expected = Time.valueOf("04:04:04");
         Mockito.when(mockResultSet.getTime(columnIndex)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Time> optionalTime = resultSet.getOptionalTime(columnIndex);
+        
+        Optional<Time> optionalTime = extendedResultSet.getOptionalTime(columnIndex);
+
         assertEquals(expected, optionalTime.orElseThrow(IllegalArgumentException::new));
     }
 
     @Test
     public void checkNullFromGetOptionalTimeByColumnIndex() throws SQLException {
         int columnIndex = 0;
-        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Time> optionalTime = resultSet.getOptionalTime(columnIndex);
+        Mockito.when(mockResultSet.getTime(columnIndex)).thenReturn(null);
+        
+        Optional<Time> optionalTime = extendedResultSet.getOptionalTime(columnIndex);
+
         assertFalse(optionalTime.isPresent());
     }
 
@@ -430,17 +686,19 @@ public class SimpleOptionalResultSetTest {
         String columnName = "ignored";
         Time expected = Time.valueOf("04:04:04");
         Mockito.when(mockResultSet.getTime(columnName)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Time> optionalTime = resultSet.getOptionalTime(columnName);
+        
+        Optional<Time> optionalTime = extendedResultSet.getOptionalTime(columnName);
+
         assertEquals(expected, optionalTime.orElseThrow(IllegalArgumentException::new));
     }
 
     @Test
     public void checkNullFromGetOptionalTimeByColumnName() throws SQLException {
         String columnName = "ignored";
-        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Time> optionalTime = resultSet.getOptionalTime(columnName);
+        Mockito.when(mockResultSet.getTime(columnName)).thenReturn(null);
+        
+        Optional<Time> optionalTime = extendedResultSet.getOptionalTime(columnName);
+
         assertFalse(optionalTime.isPresent());
     }
 
@@ -449,17 +707,19 @@ public class SimpleOptionalResultSetTest {
         int columnIndex = 0;
         Timestamp expected = Timestamp.valueOf("2004-04-04 04:04:04.004");
         Mockito.when(mockResultSet.getTimestamp(columnIndex)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Timestamp> optionalTimestamp = resultSet.getOptionalTimestamp(columnIndex);
+        
+        Optional<Timestamp> optionalTimestamp = extendedResultSet.getOptionalTimestamp(columnIndex);
+
         assertEquals(expected, optionalTimestamp.orElseThrow(IllegalArgumentException::new));
     }
 
     @Test
     public void checkNullFromGetOptionalTimestampByColumnIndex() throws SQLException {
         int columnIndex = 0;
-        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Timestamp> optionalTimestamp = resultSet.getOptionalTimestamp(columnIndex);
+        Mockito.when(mockResultSet.getTimestamp(columnIndex)).thenReturn(null);
+        
+        Optional<Timestamp> optionalTimestamp = extendedResultSet.getOptionalTimestamp(columnIndex);
+
         assertFalse(optionalTimestamp.isPresent());
     }
 
@@ -468,17 +728,19 @@ public class SimpleOptionalResultSetTest {
         String columnName = "ignored";
         Timestamp expected = Timestamp.valueOf("2004-04-04 04:04:04.004");
         Mockito.when(mockResultSet.getTimestamp(columnName)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Timestamp> optionalTimestamp = resultSet.getOptionalTimestamp(columnName);
+        
+        Optional<Timestamp> optionalTimestamp = extendedResultSet.getOptionalTimestamp(columnName);
+
         assertEquals(expected, optionalTimestamp.orElseThrow(IllegalArgumentException::new));
     }
 
     @Test
     public void checkNullFromGetOptionalTimestampByColumnName() throws SQLException {
         String columnName = "ignored";
-        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Timestamp> optionalTimestamp = resultSet.getOptionalTimestamp(columnName);
+        Mockito.when(mockResultSet.getTimestamp(columnName)).thenReturn(null);
+        
+        Optional<Timestamp> optionalTimestamp = extendedResultSet.getOptionalTimestamp(columnName);
+
         assertFalse(optionalTimestamp.isPresent());
     }
 
@@ -487,17 +749,19 @@ public class SimpleOptionalResultSetTest {
         int columnIndex = 0;
         BigDecimal expected = BigDecimal.valueOf(40_000L);
         Mockito.when(mockResultSet.getBigDecimal(columnIndex)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<BigDecimal> optionalBigDecimal = resultSet.getOptionalBigDecimal(columnIndex);
+        
+        Optional<BigDecimal> optionalBigDecimal = extendedResultSet.getOptionalBigDecimal(columnIndex);
+
         assertEquals(expected, optionalBigDecimal.orElseThrow(IllegalArgumentException::new));
     }
 
     @Test
     public void checkNullFromGetOptionalBigDecimalByColumnIndex() throws SQLException {
         int columnIndex = 0;
-        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<BigDecimal> optionalBigDecimal = resultSet.getOptionalBigDecimal(columnIndex);
+        Mockito.when(mockResultSet.getBigDecimal(columnIndex)).thenReturn(null);
+        
+        Optional<BigDecimal> optionalBigDecimal = extendedResultSet.getOptionalBigDecimal(columnIndex);
+
         assertFalse(optionalBigDecimal.isPresent());
     }
 
@@ -506,17 +770,19 @@ public class SimpleOptionalResultSetTest {
         String columnName = "ignored";
         BigDecimal expected = BigDecimal.valueOf(40_000L);
         Mockito.when(mockResultSet.getBigDecimal(columnName)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<BigDecimal> optionalBigDecimal = resultSet.getOptionalBigDecimal(columnName);
+        
+        Optional<BigDecimal> optionalBigDecimal = extendedResultSet.getOptionalBigDecimal(columnName);
+
         assertEquals(expected, optionalBigDecimal.orElseThrow(IllegalArgumentException::new));
     }
 
     @Test
     public void checkNullFromGetOptionalBigDecimalByColumnName() throws SQLException {
         String columnName = "ignored";
-        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<BigDecimal> optionalBigDecimal = resultSet.getOptionalBigDecimal(columnName);
+        Mockito.when(mockResultSet.getBigDecimal(columnName)).thenReturn(null);
+        
+        Optional<BigDecimal> optionalBigDecimal = extendedResultSet.getOptionalBigDecimal(columnName);
+
         assertFalse(optionalBigDecimal.isPresent());
     }
 
@@ -525,17 +791,19 @@ public class SimpleOptionalResultSetTest {
         int columnIndex = 0;
         Ref expected = new MockRef(40_000);
         Mockito.when(mockResultSet.getRef(columnIndex)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Ref> optionalRef = resultSet.getOptionalRef(columnIndex);
+        
+        Optional<Ref> optionalRef = extendedResultSet.getOptionalRef(columnIndex);
+
         assertEquals(expected, optionalRef.orElseThrow(IllegalArgumentException::new));
     }
 
     @Test
     public void checkNullFromGetOptionalRefByColumnIndex() throws SQLException {
         int columnIndex = 0;
-        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Ref> optionalRef = resultSet.getOptionalRef(columnIndex);
+        Mockito.when(mockResultSet.getRef(columnIndex)).thenReturn(null);
+        
+        Optional<Ref> optionalRef = extendedResultSet.getOptionalRef(columnIndex);
+
         assertFalse(optionalRef.isPresent());
     }
 
@@ -544,17 +812,19 @@ public class SimpleOptionalResultSetTest {
         String columnName = "ignored";
         Ref expected = new MockRef(40_000);
         Mockito.when(mockResultSet.getRef(columnName)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Ref> optionalRef = resultSet.getOptionalRef(columnName);
+        
+        Optional<Ref> optionalRef = extendedResultSet.getOptionalRef(columnName);
+
         assertEquals(expected, optionalRef.orElseThrow(IllegalArgumentException::new));
     }
 
     @Test
     public void checkNullFromGetOptionalRefByColumnName() throws SQLException {
         String columnName = "ignored";
-        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Ref> optionalRef = resultSet.getOptionalRef(columnName);
+        Mockito.when(mockResultSet.getRef(columnName)).thenReturn(null);
+        
+        Optional<Ref> optionalRef = extendedResultSet.getOptionalRef(columnName);
+
         assertFalse(optionalRef.isPresent());
     }
 
@@ -563,17 +833,19 @@ public class SimpleOptionalResultSetTest {
         int columnIndex = 0;
         Blob expected = new MockBlob(new byte[]{4, 0, 0, 0, 0});
         Mockito.when(mockResultSet.getBlob(columnIndex)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Blob> optionalBlob = resultSet.getOptionalBlob(columnIndex);
+        
+        Optional<Blob> optionalBlob = extendedResultSet.getOptionalBlob(columnIndex);
+
         assertEquals(expected, optionalBlob.orElseThrow(IllegalArgumentException::new));
     }
 
     @Test
     public void checkNullFromGetOptionalBlobByColumnIndex() throws SQLException {
         int columnIndex = 0;
-        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Blob> optionalBlob = resultSet.getOptionalBlob(columnIndex);
+        Mockito.when(mockResultSet.getBlob(columnIndex)).thenReturn(null);
+        
+        Optional<Blob> optionalBlob = extendedResultSet.getOptionalBlob(columnIndex);
+
         assertFalse(optionalBlob.isPresent());
     }
 
@@ -582,17 +854,19 @@ public class SimpleOptionalResultSetTest {
         String columnName = "ignored";
         Blob expected = new MockBlob(new byte[]{4, 0, 0, 0, 0});
         Mockito.when(mockResultSet.getBlob(columnName)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Blob> optionalBlob = resultSet.getOptionalBlob(columnName);
+        
+        Optional<Blob> optionalBlob = extendedResultSet.getOptionalBlob(columnName);
+
         assertEquals(expected, optionalBlob.orElseThrow(IllegalArgumentException::new));
     }
 
     @Test
     public void checkNullFromGetOptionalBlobByColumnName() throws SQLException {
         String columnName = "ignored";
-        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Blob> optionalBlob = resultSet.getOptionalBlob(columnName);
+        Mockito.when(mockResultSet.getBlob(columnName)).thenReturn(null);
+        
+        Optional<Blob> optionalBlob = extendedResultSet.getOptionalBlob(columnName);
+
         assertFalse(optionalBlob.isPresent());
     }
 
@@ -601,17 +875,19 @@ public class SimpleOptionalResultSetTest {
         int columnIndex = 0;
         Clob expected = new MockClob("forty thousand");
         Mockito.when(mockResultSet.getClob(columnIndex)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Clob> optionalClob = resultSet.getOptionalClob(columnIndex);
+        
+        Optional<Clob> optionalClob = extendedResultSet.getOptionalClob(columnIndex);
+
         assertEquals(expected, optionalClob.orElseThrow(IllegalArgumentException::new));
     }
 
     @Test
     public void checkNullFromGetOptionalClobByColumnIndex() throws SQLException {
         int columnIndex = 0;
-        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Clob> optionalClob = resultSet.getOptionalClob(columnIndex);
+        Mockito.when(mockResultSet.getClob(columnIndex)).thenReturn(null);
+        
+        Optional<Clob> optionalClob = extendedResultSet.getOptionalClob(columnIndex);
+
         assertFalse(optionalClob.isPresent());
     }
 
@@ -620,17 +896,19 @@ public class SimpleOptionalResultSetTest {
         String columnName = "ignored";
         Clob expected = new MockClob("forty thousand");
         Mockito.when(mockResultSet.getClob(columnName)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Clob> optionalClob = resultSet.getOptionalClob(columnName);
+        
+        Optional<Clob> optionalClob = extendedResultSet.getOptionalClob(columnName);
+
         assertEquals(expected, optionalClob.orElseThrow(IllegalArgumentException::new));
     }
 
     @Test
     public void checkNullFromGetOptionalClobByColumnName() throws SQLException {
         String columnName = "ignored";
-        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Clob> optionalClob = resultSet.getOptionalClob(columnName);
+        Mockito.when(mockResultSet.getClob(columnName)).thenReturn(null);
+        
+        Optional<Clob> optionalClob = extendedResultSet.getOptionalClob(columnName);
+
         assertFalse(optionalClob.isPresent());
     }
 
@@ -639,17 +917,19 @@ public class SimpleOptionalResultSetTest {
         int columnIndex = 0;
         Array expected = new MockArray("forty thousand");
         Mockito.when(mockResultSet.getArray(columnIndex)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Array> optionalArray = resultSet.getOptionalArray(columnIndex);
+        
+        Optional<Array> optionalArray = extendedResultSet.getOptionalArray(columnIndex);
+
         assertEquals(expected, optionalArray.orElseThrow(IllegalArgumentException::new));
     }
 
     @Test
     public void checkNullFromGetOptionalArrayByColumnIndex() throws SQLException {
         int columnIndex = 0;
-        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Array> optionalArray = resultSet.getOptionalArray(columnIndex);
+        Mockito.when(mockResultSet.getArray(columnIndex)).thenReturn(null);
+        
+        Optional<Array> optionalArray = extendedResultSet.getOptionalArray(columnIndex);
+
         assertFalse(optionalArray.isPresent());
     }
 
@@ -658,17 +938,19 @@ public class SimpleOptionalResultSetTest {
         String columnName = "ignored";
         Array expected = new MockArray("forty thousand");
         Mockito.when(mockResultSet.getArray(columnName)).thenReturn(expected);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Array> optionalArray = resultSet.getOptionalArray(columnName);
+        
+        Optional<Array> optionalArray = extendedResultSet.getOptionalArray(columnName);
+
         assertEquals(expected, optionalArray.orElseThrow(IllegalArgumentException::new));
     }
 
     @Test
     public void checkNullFromGetOptionalArrayByColumnName() throws SQLException {
         String columnName = "ignored";
-        Mockito.when(mockResultSet.wasNull()).thenReturn(true);
-        OptionalResultSet resultSet = ResultSetUtils.optional(mockResultSet);
-        Optional<Array> optionalArray = resultSet.getOptionalArray(columnName);
+        Mockito.when(mockResultSet.getArray(columnName)).thenReturn(null);
+        
+        Optional<Array> optionalArray = extendedResultSet.getOptionalArray(columnName);
+
         assertFalse(optionalArray.isPresent());
     }
 }
